feat: move gpp to InternalUserConsentDataManager
