+++ PrebidMobile/PrebidMobileRendering/Networking/Parameters/PBMBasicParameterBuilder.m
@@ -17,6 +17,8 @@
 #import "PBMMacros.h"
 #import "PBMORTB.h"
 
+#import "InternalUserConsentDataManager.h"
+
 #import "PBMBasicParameterBuilder.h"
 
 #import "PrebidMobileSwiftHeaders.h"
@@ -119,8 +121,8 @@
     
     bidRequest.regs.coppa = self.targeting.coppa;
     bidRequest.regs.ext[@"gdpr"] = [self.targeting getSubjectToGDPR];
-    bidRequest.regs.ext[@"gpp"] = [self.targeting gppHDRString];
-    bidRequest.regs.ext[@"gpp_sid"] = [self.targeting gppSID];
+    bidRequest.regs.ext[@"gpp"] = InternalUserConsentDataManager.gppHDRString;
+    bidRequest.regs.ext[@"gpp_sid"] = InternalUserConsentDataManager.gppSID;
     
     [self appendFormatSpecificParametersForRequest:bidRequest];
 }
